+++ Rubrik/Private/Get-RubrikAPIData.ps1
@@ -1,7 +1,11 @@
-<#
-    Helper function to retrieve API data from Rubrik
-#>
 function Get-RubrikAPIData($endpoint) {
+    <#
+        .SYNOPSIS
+        Helper function to retrieve API data from Rubrik
+
+        .DESCRIPTION
+        Function which defines the structure of each cmdlet's API request and response based on the software version the connected Rubrik cluster.
+    #>
     $api = @{
         Example                        = @{
             '1.0' = @{